move lottie.json animation to a common animation directory
