+++ apps/ledger-live-mobile/src/animations/lottie.json
