+++ apps/ledger-live-mobile/src/newArch/features/WalletSync/screens/Activation/ActivationLoading.tsx
@@ -8,7 +8,7 @@ import { AnalyticsPage } from "../../hooks/useLedgerSyncAnalytics";
 import GradientContainer from "~/components/GradientContainer";
 import Animation from "~/components/Animation";
 import { Flex, Text } from "@ledgerhq/native-ui";
-import lottie from "~/screens/ReceiveFunds/assets/lottie.json";
+import lottie from "~/animations/lottie.json";
 import { useTheme } from "styled-components/native";
 import { useTranslation } from "react-i18next";
 import { useDispatch, useSelector } from "react-redux";
